+++ lib/src/curl_options.dart
@@ -1,8 +1,6 @@
 import 'package:colored_logger/colored_logger.dart';
 import 'package:dio_curl_interceptor/src/types.dart';
 
-import 'curl_formatters.dart';
-
 class CurlOptions {
   const CurlOptions({
     this.status = true,
@@ -11,16 +9,10 @@ class CurlOptions {
     this.onRequest = const RequestDetails(),
     this.onResponse = const ResponseDetails(),
     this.onError = const ErrorDetails(),
-    this.formatter,
     this.behavior = CurlBehavior.simultaneous,
     this.printer,
-    this.disabledSuggestions = false,
-    this.colored = true,
+    this.prettyConfig = const PrettyConfig(),
   });
-
-  factory CurlOptions.escapeNewlinesString() => CurlOptions(
-        formatter: CurlFormatters.escapeNewlinesString,
-      );
 
   /// Show status code, status name, method, uri, response time
   final bool status;
@@ -34,17 +26,12 @@ class CurlOptions {
   /// Define behavior of curl how it's printed
   final CurlBehavior? behavior;
 
-  /// Used to format response body
-  final String Function(dynamic body)? formatter;
-
-  /// Show colored output, set to false if you want to disable colored output.
-  final bool colored;
-
   /// The printer function for printing the curl command.
   final void Function(String text)? printer;
 
-  /// Disable suggestions, set to true if you want to disable all suggestions setup.
-  final bool disabledSuggestions;
+  /// Configuration for pretty printing HTTP requests and responses.
+  /// Controls the visual appearance of the output when pretty printing is enabled.
+  final PrettyConfig prettyConfig;
 
   final RequestDetails? onRequest;
   final ResponseDetails? onResponse;
@@ -92,4 +79,37 @@ class ErrorDetails extends CurlDetails {
   });
 
   final bool responseBody;
+}
+
+/// Configuration options for pretty printing HTTP requests and responses in a box format.
+class PrettyConfig {
+  const PrettyConfig({
+    this.blockEnabled = true,
+    this.useUnicode = true,
+    this.lineLength = 80,
+    this.disabledSuggestions = false,
+    this.colorEnabled = true,
+    this.emojiEnabled = true,
+  });
+
+  /// Enable pretty printing of HTTP requests and responses in a box format.
+  /// When enabled, the output will be formatted in a visually appealing box.
+  final bool blockEnabled;
+
+  /// Show colored output, set to false if you want to disable colored output.
+  final bool colorEnabled;
+
+  /// Show emoji output, set to false if you want to disable emoji output.
+  final bool emojiEnabled;
+
+  /// Disable suggestions, set to true if you want to disable all suggestions setup.
+  final bool disabledSuggestions;
+
+  /// Use Unicode box-drawing characters for pretty printing.
+  /// When true, uses Unicode characters like ╔, ╗, ╚, ╝, ═, ║.
+  /// When false, uses ASCII characters like +, |, =.
+  final bool useUnicode;
+
+  /// Length of separator lines in pretty printing.
+  final int lineLength;
 }