+++ lib/src/dio_curl_interceptor.dart
@@ -1,7 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:dio_curl_interceptor/dio_curl_interceptor.dart';
 
-export 'curl_formatters.dart';
 export 'curl_helpers.dart';
 export 'curl_options.dart';
 export 'curl_utils.dart';
@@ -20,7 +19,7 @@ String? genCurl(RequestOptions options, {bool convertFormData = false}) {
     return CurlHelpers.generateCurlFromRequestOptions(options);
   } catch (e) {
     ColoredLogger.info(
-        '[CurlInterceptor] Unable to create a CURL representation to ${options.uri.toString()}');
+        '[CurlInterceptor] Unable to create a cURL representation to ${options.uri.toString()}');
     return null;
   }
 }