+++ lib/src/emoji.dart
@@ -20,25 +20,21 @@ class Emoji {
   static const String patch = '📝'; // PATCH
   static const String delete = '🗑️'; // DELETE
 
-  // others
-  static const String lightBulb = '💡'; // Light bulb
-  static const String pin = '📌'; // Pin
+  // misc
+  static const String package = '📦'; // Package
+  static const String link = '🔗'; // Link
+  static const String document = '🧾'; // Document
+  static const String image = '🖼️'; // Image
+  static const String audio = '🔊'; // Audio
+  static const String video = '📹'; // Video
+  static const String folder = '📁'; // Folder
+  static const String database = '🗃️'; // Database
+  static const String cloud = '☁️'; // Cloud
+  static const String star = '⭐️'; // Star
   static const String gear = '⚙️'; // Gear
-  static const String wrench = '🔧'; // Wrench
-  static const String hammer = '🔨'; // Hammer
+  static const String pin = '📌'; // Pin
+  static const String lightBulb = '💡'; // Light bulb
   static const String lock = '🔒'; // Lock
   static const String key = '🔑'; // Key
   static const String tag = '🏷️'; // Tag
-  static const String tagPlus = '🏷️+'; // Tag plus
-  static const String tagMinus = '🏷️-'; // Tag minus
-  static const String arrowRight = '➡️'; // Arrow right
-  static const String arrowLeft = '⬅️'; // Arrow left
-  static const String arrowUp = '⬆️'; // Arrow up
-  static const String arrowDown = '⬇️'; // Arrow down
-  static const String plus = '➕'; // Plus
-  static const String minus = '➖'; // Minus
-  static const String star = '⭐️'; // Star
-  static const String heart = '❤️'; // Heart
-  static const String thumbsUp = '👍'; // Thumbs up
-  static const String thumbsDown = '👎'; // Thumbs down
 }